feat: Overhaul project with modern tooling and new features
This commit introduces a comprehensive set of improvements to the Cosmic Wormhole Experience project, addressing the user's request to enhance its efficiency, realism, and code quality.

The key changes include:

1.  **Modern Development Environment:**
    - The project has been refactored to use Vite as a build tool and npm for dependency management.
    - All JavaScript code has been migrated to ES6 modules, with dependencies like 'three' now managed via `package.json`.
    - CDN links have been removed in favor of local, version-controlled dependencies.

2.  **Performance Optimization:**
    - Procedural client-side texture generation has been replaced with a static asset loading system. This significantly reduces initial load time and CPU overhead for the user.
    - Placeholder textures have been added to `/public/assets`.

3.  **Graphics Enhancement:**
    - A new, animated accretion disk has been added around the wormhole's mouth.
    - This effect is created using a custom GLSL shader, adding a dynamic and visually impressive element to the scene.

4.  **Accessibility Improvements:**
    - ARIA roles have been added to the main UI panels to provide better semantic structure for screen readers.
    - CSS has been updated to ensure all interactive elements have clear, visible focus states for improved keyboard navigation.
    - Problematic `outline: none` styles have been removed.

These changes make the project more robust, performant, maintainable, and accessible, aligning it with modern web development standards.
