+++ index.html
@@ -27,8 +27,8 @@
   <!-- Interface Panels -->
   <div class="interface">
     <!-- Information Panel -->
-    <div class="info-panel panel">
-      <h3>Einstein-Rosen Bridge</h3>
+    <div class="info-panel panel" role="region" aria-labelledby="info-panel-title">
+      <h3 id="info-panel-title">Einstein-Rosen Bridge</h3>
       <p>A theoretical passage through spacetime creating a shortcut connecting distant regions of the universe.</p>
       <p>This visualization represents a stable traversable wormhole with modified quantum energy conditions.</p>
       <div class="wormhole-types">
@@ -44,8 +44,8 @@
     </div>
 
     <!-- Controls Panel -->
-    <div class="controls-panel panel">
-      <h3>Navigation Controls</h3>
+    <div class="controls-panel panel" role="region" aria-labelledby="controls-panel-title">
+      <h3 id="controls-panel-title">Navigation Controls</h3>
       <div class="control-group">
         <label for="distortion">Space Distortion: <span id="distortion-value">50%</span></label>
         <input type="range" id="distortion" min="0" max="100" value="50" aria-valuemin="0" aria-valuemax="100" aria-valuenow="50" />
@@ -62,8 +62,8 @@
     </div>
 
     <!-- Scientific Panel -->
-    <div class="scientific-panel panel">
-      <h3>Scientific Data</h3>
+    <div class="scientific-panel panel" role="region" aria-labelledby="scientific-panel-title">
+      <h3 id="scientific-panel-title">Scientific Data</h3>
       <div class="data-display">
         <div class="data-item">
           <span class="data-label">Throat Diameter:</span>
@@ -102,57 +102,6 @@
     <div class="fps-counter" id="fps-counter">60 FPS</div>
   </div>
 
-  <script src="https://cdnjs.cloudflare.com/ajax/libs/three.js/r128/three.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/controls/OrbitControls.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/postprocessing/EffectComposer.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/postprocessing/RenderPass.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/postprocessing/ShaderPass.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/shaders/CopyShader.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/shaders/LuminosityHighPassShader.min.js"></script>
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/postprocessing/UnrealBloomPass.min.js"></script>
-  <!-- Important: GlitchPass must be included since it's used in setup.js -->
-  <script src="https://cdn.jsdelivr.net/npm/three@0.128.0/examples/js/postprocessing/GlitchPass.min.js"></script>
-  
-  <!-- Handle modules properly -->
-  <script type="module">
-    // Import modules
-    import { initScene } from './setup.js';
-    import { initAnimation } from './animation.js';
-    
-    // Add specific error handler for script loading errors
-    window.addEventListener('error', function(event) {
-      console.error('Script error:', event.error);
-      const loadingStatus = document.getElementById('loading-status');
-      if (loadingStatus) {
-        loadingStatus.textContent = "Error loading wormhole: " + (event.error ? event.error.message : "Unknown error");
-      }
-    }, { once: false, capture: true });
-    
-    // Initialize application after DOM is ready
-    document.addEventListener('DOMContentLoaded', async () => {
-      try {
-        // Initialize scene and wait for completion
-        await initScene();
-        // Start animation loop
-        initAnimation();
-      } catch (error) {
-        console.error('Initialization error:', error);
-        
-        // Update loading status with error message
-        const loadingStatus = document.getElementById('loading-status');
-        if (loadingStatus) {
-          loadingStatus.textContent = "Error initializing wormhole: " + error.message;
-        }
-        
-        // Optionally hide loading screen after error
-        setTimeout(() => {
-          const loadingScreen = document.getElementById('loading');
-          if (loadingScreen) {
-            loadingScreen.classList.add('hidden');
-          }
-        }, 300);
-      }
-    });
-  </script>
+  <script type="module" src="/src/main.js"></script>
 </body>
 </html>
