+++ style.css
@@ -232,7 +232,8 @@
    transition: all 0.3s ease;
  }
  
- input[type="range"]::-webkit-slider-thumb:hover {
+ input[type="range"]::-webkit-slider-thumb:hover,
+ input[type="range"]:focus-visible::-webkit-slider-thumb {
    background: var(--color-accent);
    box-shadow: 0 0 15px rgba(0, 220, 255, 0.7);
    transform: scale(1.1);
@@ -249,6 +250,12 @@
    transition: all 0.3s ease;
  }
  
+ input[type="range"]:focus-visible::-moz-range-thumb {
+  background: var(--color-accent);
+  box-shadow: 0 0 15px rgba(0, 220, 255, 0.7);
+  transform: scale(1.1);
+}
+
  input[type="range"]::-moz-range-track {
    background: linear-gradient(to right, 
      rgba(0, 100, 200, 0.3) 0%, 
@@ -275,9 +282,14 @@
    transition: all 0.3s ease;
  }
  
- .select-control:hover {
+ .select-control:hover, .select-control:focus-visible {
    border-color: rgba(0, 220, 255, 0.8);
    box-shadow: 0 0 15px rgba(0, 150, 255, 0.3);
+ }
+
+ .select-control:focus-visible {
+  outline: 2px solid var(--color-primary);
+  outline-offset: 2px;
  }
  
  /* Enhanced journey button with animated shine effect */
@@ -297,7 +309,6 @@
    font-size: 0.85rem;
    position: relative;
    overflow: hidden;
-   outline: none;
    box-shadow: 0 4px 15px rgba(0, 100, 200, 0.3);
  }
  
@@ -328,6 +339,14 @@
    transform: translateY(1px);
    box-shadow: 0 2px 10px rgba(0, 150, 255, 0.3);
  }
+
+ .journey-btn:focus-visible {
+  background: linear-gradient(135deg, rgba(0, 180, 255, 0.5), rgba(0, 140, 255, 0.6));
+  box-shadow: 0 5px 20px rgba(0, 150, 255, 0.5);
+  border-color: rgba(0, 220, 255, 0.7);
+  outline: 2px solid var(--color-primary);
+  outline-offset: 2px;
+}
  
  /* Enhanced data display */
  .data-display {
